Create ConvertTo-PowerShellArray.ps1
Updated version of an function that only worked with PowerShell ISE to work with VS Code by pulling the InputList from the clipboard or a quoted string on the commandline. Improvements include removing blank lines and a generalized sort switch with regex detection of IPv4 addresses; IP addresses are sorted using the [version] class.