Major Overhaul
Extensive re-write by Claude.AI and my friend Joe. After upgrading to Windows 11 the script was not reliably parsing many of the inputs from the clipboard that should have been evaluated as CRLF for multiline strings. It was especially bad when using Active Directory Distinguished Names. 

Using $(Get-Clipboard -Raw) was a key change